requirements.txt
asttokens==2.4.1
attrs==23.2.0
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.2
cssselect==1.2.0
decorator==5.1.1
executing==2.0.1
filelock==3.13.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.1
jmespath==1.0.1
json5==0.9.14
lxml==5.1.0
matplotlib-inline==0.1.6
packaging==23.2
parsel==1.8.1
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
Protego==0.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
Pygments==2.17.2
pyOpenSSL==24.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
queuelib==1.6.2
requests==2.31.0
requests-file==2.0.0
scrapeops_scrapy==0.5.4
Scrapy==2.11.0
service-identity==24.1.0
six==1.16.0
stack-data==0.6.3
tld==0.13
tldextract==5.1.1
traitlets==5.14.1
Twisted==22.10.0
typing_extensions==4.9.0
urllib3==2.2.0
w3lib==2.1.2
wcwidth==0.2.13
zope.interface==6.1
